neuro-ai-backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-migrate
werkzeug
python-dotenv
transformers
torch
tqdm
flask-cors
PyPDF2
safetensors